requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
gspread==5.12.4
python-dotenv==1.0.0
